Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/012-expected.txt


FAIL Link with onclick navigation to javascript url with delayed document.write and href navigation  assert_equals: expected "href" but got "write"
 Test write
